Don't automatically unstar changes
The original idea was to unstar a change once we it is either merged or
abandoned. This means that the next time we sync the change will be
removed from Trello. It turned out that I like to manually remove the
stars so that I can see the resolution.

This may come back in the future.